% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/<LearnRoxygen>.R
\docType{package}
\name{LearnRoxygen}
\alias{LearnRoxygen}
\alias{LearnRoxygen-package}
\title{LearnRoxygen: A package for understanding the roxygen2 package.}
\description{
The LearnRoxygen package provides many functions!
}
\section{Data parsing functions}{

Functions to parse data ...
}

\section{Data aggregating functions}{

Functions to aggregate data ...
}

